requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[aiosqlite]
python-jose
passlib[bcrypt]
python-dotenv
requests
python-multipart
aiofiles
b2sdk
pytest
httpx
pytest-asyncio
pytest-mock
pyfakefs